backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
xgboost
shap
mlflow
python-dotenv
httpx
